Remove unit tests for no-op filter in convolutions
Unit tests loop across the whole list of filters for each filter type,
which includes the no-op filter when sub_x or sub_y = 0. In practice,
however, the convolution functions are never called for such filters
(convolve_copy will be used instead), so remove these test cases. The
code accommodating this special no-op filter in the Neon paths will be
removed in a subsequent patch.

Change-Id: Ib363f15b35b1bdec796e4e8f490b9fae45328f54