Add SVE2 implementation for 12-tap av1_convolve_2d_sr
The vertical pass of av1_convolve_2d_sr operates on 16-bit elements, so
we can use the SVE 16-bit dot-product to accelerate it. Given that the
presence of SVE2 implies the presence of Neon_I8MM, keep the i8mm
implementation for the horizontal pass. This gives an uplift of up to
25% compared to the full i8mm version.

Change-Id: I909ed998f83593ce9d0108c8e2a41f71c08a73de