Enable I8MM when compiling SVE2 files
The CPU feature detection logic assumes that if SVE2 is available then
all previous features are available as well (SVE, NEON_DOTPROD and
NEON_I8MM). Add 'i8mm' to the compiler flags for SVE2 files to reflect
these constraints. It's not necessary to add 'sve' or 'dotprod' as they
are implied by the 'armv9' flag.

Change-Id: I7e81b041910676af870eb0f45d3ed3a4014d31be