Add Neon DotProd implementation for av1_convolve_2d_scale
Add an Armv8.4 DotProd implementation for the horizontal filtering
part of av1_convolve_2d_scale.

The vertical pass operates on 16-bit types so the Armv8.0
implementation must still be used. Move functions associated with the
vertical pass into a header file so they can be shared by both the
Armv8.0 and Armv8.4 DotProd paths.

Add the corresponding tests as well.

Change-Id: I4865953a1aff7bde9b786d65a140da2501894cc7