Reduce encode_time on slide change occurs on non_reference
For rtc-screen: for slide changes that occur on a non_reference frame
(i.e, top temporal enhancement layer frames): add a speed feature to
skip encode the frame. This option is currently set as default.

Regardless of this option: adjust some speed features and turn off
palette to significantly reduce the encode_time spike for that frame,
if it's not skip encoded.

Change-Id: I70cefdfa7d158fd53e8c4720c7266fdd24fe9635