rtc: Remove logic to set sb_size for dynamic mode
For real-time and screen mode: when the --sb-size=dynamic
is used (which is default), the sb_size was being set to 64 under certain conditions (based on resolution and #threads). This
selection of 64 size is causing the regression (compared to 1
thread which is using 128), as shown in bug attached below.

Remove this internal logic for sb_size switch in dynamic mode
for now. Will follow-up to investigate how 64 vs 128 caused
such significant regression in those clips.

Bug: b/343429036

Change-Id: Id118f587c782d5e21ccaa7cb73355b86bb0267bb