rtc: Remove logic to set sb_size for dynamic mode
For real-time and screen mode: when the --sb-size=dynamic
is used, the sb_size was being set to 64 under certain
conditions (based on resolution and #threads). This selection
of 64 size is causing the big regression (compared to 1 thread
which is using 128) as shown in this issue:
https://buganizer.corp.google.com/issues/343429036

Removing this logic for now, and leave it to the user
to select set the --sb-size= 64 or 128.

Change-Id: Id118f587c782d5e21ccaa7cb73355b86bb0267bb