rtc: Condition QP adjustment on rc->q_1/2_frame > 0
This fix only affects the initial QPs after very first
frame for TL enhancement frames in CBR mode.
Move the check that rc->q_1/2_frame is set to the top
of the adjustment logic, otherwise if there are dropped
frames after the very first frame, the QP can oscillate on
temporal enhancement frames.

Bug: b/343429192
Change-Id: I121934c2c7e7522e4e40762191ed3005437841c2