Add Neon I8MM horiz 2x1 scale spec. impl for convolve_2d_scale
AV1 has a limit on the scale ratio, specifically, the reference
resolution cannot be more than 2 times the source resolution in any
dimension. Given that the algorithm uses higher precision
(1/1024-pel) for the step size (chapter 7.11.3.4. [1]), the
horizontal scaling function can be easily optimised for this specific
case. The indices of the source pixel to be interpolated are
calculated using the (subpel_qn + x * step) >> 1024 equation, which
can be simplified if step is a multiple of 1024.

Add implementation that specialises on x_step_qn equals to 2048, that
gives an uplift of around 28% when a 2x1 scaling is applied.

[1]https://aomediacodec.github.io/av1-spec/av1-spec.pdf

Change-Id: I87b0730d0a75c534813f154555cbdb473b445438