av1_cx_iface: Do no require timebase.num < timebase.den
This has been in libvpx for a while now.
VP9: https://chromium-review.googlesource.com/c/webm/libvpx/+/332449
VP8: https://chromium-review.googlesource.com/c/webm/libvpx/+/274107

Generally, we set the timebase to be the inverse of the desired frame
rate. So this change allows us to express that using timebase (for
e.g. a frame rate of 0.5 fps will be 2/1).

Tested with an input video of 0.5 fps and the output is as intended
with this change (it fails before this change).

Change-Id: I296620f7625824ff460aa503388ed821c7e6ddd0