av1_cx_iface: Allow sub-second frame rates
This has been in libvpx for a while now.
VP9: https://chromium-review.googlesource.com/c/webm/libvpx/+/332449
VP8: https://chromium-review.googlesource.com/c/webm/libvpx/+/274107

This allows us to encode frames at a rate lower than 1 fps.

Tested with an input video of 0.5 fps and the output is as intended
with this change (it fails before this change).

Change-Id: I296620f7625824ff460aa503388ed821c7e6ddd0