rtc: Fix multi-threading settings for SVC datarate tests
Pass in tile_column_ and tile_row_ from each test with
multi-threading. This fixes the 4 thread test case where
the intention was to use 2x2 (tile_col x tile_row) config,
the current test was not doing that.

Also a add test for (1SL, 2TL) with 4 threads for more
coverage.

Change-Id: I910f290ad3705cfcee8b9a874e13e911243a3f54