rtc: Add decimation factors to frame_dropper
In current frame_dropper there is only one
decimation_factor (=1), which drops every other frame,
and only when buffer is above 0 (when below 0 the frame
was always dropped).

Add option (enable_decimation_factors) for different
framerate decimation factors (1, 2, 3, ..), with condition
based on buffer_level to select the decimation factor.
Use larger factor when buffer is below 0.

WIP: to be tested

Change-Id: I1c30ea85dd8ee9beb17c020a52af85bc26aa6f32