rtc: Add postencode frame drop feature
Allows for dropping a frame after encoding,
based on buffer_level and encoded frame size.
If decision is to drop: a flag is set (rc->force_max_q)
to force next incoming frame to be encoded at max_q.

Only used for RTC and when the frame dropper is on
(drop_frames_water_mark is above 0). Disable usage
for the external RC.

Feature is off by default, added control to enable it.

Change-Id: If55df612cf8fec630f753dd11b796a7c9673ddd8