Fix active_worst_quality after keyframe for temporal layers
For temporal enhancement layers in RTC mode:
for the active_worst_quality update after a key frame:
use the avg_qindex_key for the ambient_qp.

This reduces some sawtooth-like behavoior in psnr/qp
observed for temporal layers with periodic keyframes.

This also exposed a bug in the external RC where
frames_since_key was not updated in the correct place.

Change-Id: Ia63819a98823b13aefdde79b934fb30b78953986