rtc: Adjust RC for keyframes based on source content
For keyframes in rtc: compute spatial metrics
and use it, along with the last keyframe
encoded/target_size to adjust the rc_bits_per_mb.

This feature is used to avoid overshoot when
max_intra_max_intra_bitrate_pct is enabled and
has small values. Feature is off for now.

Change-Id: I0316d9303cb0ada4ed75948bbfc8196218870514