Fix keyframe counter update for dropped frames.
The counters are always updated in update_rc_counters(),
for encoded or dropped frames. So the additional update
in postencode_update_drop_frame() is not needed.

Remove the update in postencode_update_drop_frame(). And
this then requires an update in RateControlRTC::ComputeQP(),
to keep external and internal RC the same.

Change-Id: I2f1f67f8aa6586603801d5d65889445e2000d720