Fix double linking with -lvmaf
In the linker command line for libaom.so.3.9.1, I see
  -L/home/wtc/tmp/prefix/lib/x86_64-linux-gnu  -lvmaf  -lvmaf
So we link with -lvmaf twice.

The relevant variables in CMakeCache.txt are:
  VMAF_LDFLAGS:INTERNAL=-L/home/wtc/tmp/prefix/lib/x86_64-linux-gnu;-lvmaf
  VMAF_LDFLAGS_OTHER:INTERNAL=
  VMAF_LIBDIR:INTERNAL=/home/wtc/tmp/prefix/lib/x86_64-linux-gnu
  VMAF_LIBRARIES:INTERNAL=vmaf
  VMAF_LIBRARY_DIRS:INTERNAL=/home/wtc/tmp/prefix/lib/x86_64-linux-gnu

So passing ${VMAF_LDFLAGS} to target_link_libraries() should suffice. It
is not necessary to also pass ${VMAF_LIBRARIES}.

Alternatively, we could pass ${VMAF_LIBRARY_DIRS} ${VMAF_LIBRARIES}.

Change-Id: I2a42a6c34d69452a42d936138f0971eeb956b0ed