Fix stack overflow on Arm NEON
av1 selfguided restoration uses a couple of large temporary
arrays with the total size up to 95kB on stack.
We should either allocate large data on heap (which may have
performance impact on decoding) or make it static within the function
scope. The latter seems to be the least intrusive and performance friendly solution.

BUG=aomedia:355499165

Change-Id: I0f98693928d2b4f4e750a4adef7137a8b774d6cd