Add unit tests for av1_resize_and_extend_frame
Add unit tests for the Neon and SSSE3 implementation of
av1_resize_and_extend_frame.

The implementation is a port of libvpx unit tests[1] for
vp9_scale_and_extend_frame.

Unit tests for 4 to 3 scaling are disabled due to a bug[2] in the
Neon and SSSE3 implementation.

Unit tests for SSSE3 implementation are disabled due to a bug[2].

[1] https://source.chromium.org/chromium/chromium/src/+/main:third_party/libvpx/source/libvpx/test/vp9_scale_test.cc
[2] Bug: 363916152

Change-Id: I128adce8834f8835cb7992b424f7b9d39d2bbc28