Change default_extra_cfg to an array
Add the `usage` field to the beginning of struct av1_extracfg. It has
the same meaning as the g_usage field in aom_codec_enc_cfg_t.

default_extra_cfg[0] can be used independent of the `usage` value. In
the future a client of libaom can request a member of the
default_extra_cfg array for a given `usage`.

Bug: 359341194
Change-Id: Idb90d7843b3122f0ac0c5f07216c88d776ebef27