rtc: Fix for artifacts for screen with active_maps
For active_maps: remove the logic to disable cdef
at frame level based on percent_blocks_inactive, as
this reduces the artifacts reported in issue below.

The cdef can be disabled on block level for rt mode,
and this seems sufficient for speeed/quality gains,
so remove the frame-level disabling.

Bug: 42302583
Change-Id: I23669b208f45dad5ae44556166fd6c25c45a382b