Optimization of the Multithreading Preparation Phase
The prepare_enc_workers() costs so much, so we move the thread
preparations to the threads themselves. We get 58%(179%/113%)
improvement when core_num=64.

With using following command we can see the improvement
as below on icelake:
./aomenc   --threads=${core_num}  -o out.av1  --tile-columns=4 --tile-rows=2 --row-mt=1 --fp-mt=1   --cpu-used=11 --rt Bosphorus_1920x1080_120fps_420_8bit_YUV.y4m

Bosphorus_1920x1080_120fps_420_8bit_YUV.y4m is from
http://ultravideo.cs.tut.fi/video/Bosphorus_1920x1080_120fps_420_8bit_YUV_Y4M.7z

Take the score(fps) of main branch with 8 cores as base

core_num	8	16	24	32	40	48	56	64
original	100%	127%	130%	120%	116%	117%	115%	113%
optimized	101%	141%	160%	161%	163%	177%	177%	179%

Change-Id: Iecf5017d0f631d92f25935b74cec533045b1ecfe