Set palette to 1 for default_extra_cfg for CONFIG_REALTIME_ONLY
For RTC CONFIG_REALTIME_ONLY build: enable palette mode
as preset. For video mode this has no effect (palette is
only used for screen content for rtc), but for screen content
palette has big effect, so better to keep it on default.

This reverts it back to the behavior before the commit:
https://aomedia-review.googlesource.com/c/aom/+/192786

Change-Id: I2401b125d7b6dfb882141256dd8efd4404d98f09