rtc: Fix setting of intra-only frame for set_ref_frame_config
When setting the reference structure with
SET_REF_FRAME_CONFIG: if no references are set
for delta frames then that frame should be set internally
to an intra-only frame. But this check was only being
done if layers are used (number_temporal > 1 or number_spatial > 1)
(i.e., use_svc is set). So single layer was not being checked.

Fix here is to check and set to intra-only regardless of layers.

Added unittest, based on:
https://chromium-review.googlesource.com/c/chromium/src/+/5809671

Bug: 363016123
Change-Id: Ie9d9ce190dc407a50e592a5f8bf47c2a26779e9b