Optimise av1_apply_temporal_filter_neon_dotprod
To perform filtering we have to calculate sum of square of elements
which are part of a 5x5 size window. This window is a sliding window,
which means, that in a new loop cycle we are calculating the result of
4 rows from the loop cycle one before and a newly loaded row. Instead
of recalculating the dot product at every loop cycle, subtract the
sum of square of the elements of the row which is not part of the
sliding window anymore and add the sum of elements of the newly
loaded row.

Vectorise the addition between the sum of squares of the sliding
windows and luma_sse_sum.

Change-Id: Ib5f251161f06174ca129cc275745ca813e09c5b0