rtc: Allow QP to decrease more aggressively for static content
For RTC static content, with stable buffer level, relax the
limit on how much QP can change frame-to-frame and allow for
a delta bias. This allows for faster quality ramp-up and QP
convergence on static content following scene change.
Added feature to control it, enabled for screen content at
the higher speed settings.

Bug: 362936830
Change-Id: I751ebde7295b1e70851aeec00406674fa112d2e6