warp_plane_{neon_i8mm,sve}.c: Move add_const into USDOT
The existing code performs an addition after the dot-product
instructions, however it is possible to instead initialize the
dot-product accumulator with a non-zero value and perform the
accumulation there.

This doesn't actually save an instruction since the compiler must emit a
MOV to set up an accumulator instead of an ADD, however this still gives
a win overall since it shortens the length of the overall dependency
chain.

Benchmarking on a Neoverse V2 machine with LLVM 18 this gives a geomean
0.5% reduction in the times reported by the warp affine speed tests,
including a 2% geomean reduction for the most common alpha=0 beta=0
cases.

Change-Id: Iae5d3b37764628cf4c90dc15a124129b41d2953b