Check it's safe to cast 64bit duration to uint32_t
In encoder_encode(), check it is safe to cast the return value of
ticks_to_timebase_units() (an int64_t duration value) to uint32_t.

Note that this cast may be safe because aom_codec_encode() requires its
duration parameter be <= UINT32_MAX. But without an analysis to prove
that the return value of this ticks_to_timebase_units() call must
be <= UINT32_MAX, it is safer to check that.

Change-Id: I6c6d3e8ff0471f4413b917d451eaab086a2c26db