rtc-svc: Reset ref_map_idx for references not used
For layered rtc coding, with the set_ref_frame_config:
for a reference not used reset its ref_map_idx
to the first reference used. This is needed for the
issue attached below.

This change has a bugfix for dropped frames regarding
the svc_update_buffer_slot_refreshed. Also updates a
setting in rtc unittest.

Bug: 362123224

Change-Id: I39a0e57b5d497b60179be4eeaed3cab924a493c4