Optimize Neon impl. of av1_compute_stats with no downsampling
Use the same implementation as the high bitdepth version for the
non-downsampling case. Unfortunately downsampling means we cannot reuse
as much computation, thus applying the same kind of implementation would
not be as effective.

Since the new Neon non-downsampling implementation is faster than the
SVE implementation, only use the SVE implementation for the downsampling
case.

Change-Id: I3df980b43a2ce43fb9409a227d663ea2c0deb143