Some cleanup for av1_convert_sect5obus_to_annexb()
Delete the variable total_bytes_read. The sum of total_bytes_read and
remaining_size is a constant (== *frame_size), so we only need one of
them.

Change two memmove() calls to memcpy().

Use the &= operator to simplify an assignment.

Subtract obu_header_size from the second argument to aom_uleb_decode()
because obu_header_size is the offset in the first argument.

Move the variable obu_size earlier to eliminate an occurrence of the
expression obu_header_size + obu_payload_size.

Bug: 42302568, 361552711
Change-Id: I9ad6527546bb22b15339a23b402ba2a672c86f24