rtc-svc: Fix to reset ref_idx for svc
Constrain the reset only for the single reference
case for now. This is causing failures in WebRTC
with the latest libaom.

The failing tests are doing spatial layers with LAST
as the spatial/inter-layer reference, which may be the
source of the issue. Will follow-up with unitest and
more investigation.

Bug: 369633254
Change-Id: I92ea2eb90bb1ea2433ab6d355f895aa737089b1d