rtc-svc: Fix to reset ref_idx for svc
The rtc_ref->ref_idx[] is mapped to the
remapped_idx[] at the start of encoding, so on
reset of ref_idx in we should also reset the
remapped_ref_idx, as this is used further below
in the function write_frame_size_with_refs(line ~3151).

The previous fix of constraining the reset
to single reference is not quite correct and
can still fail.

Added two test cases that trigger the failure.
Bug: 369633254
Change-Id: I86490e3f96ffcea7fb6e972a7d2085675c65fcc4

Change-Id: Ifcc48edff86a2eca6253ea00c81f8901c8ff0631