rtc: Constrain the reset of ref_idx for spatial layers
Constrain the rtc_ref->ref_idx[] reset for spatial layers,
as this can cause decode failures when LAST is used as
the spatial reference. This is because the rtc_ref->ref_idx[]
is mapped to the remapped_idx[] at the start of encoding,
and this is used further below (after the reset) in the function
write_frame_size_with_refs(line ~3151).

An alternate fix would be to reset
cm->remapped_ref_idx[ref_frame - 1] = map_idx
after line 3125. But for now we take the simpler fix
of constraining the reset.

The previous fix of constraining the reset
to single reference is not quite correct and
can still fail for spatial layers. The previous fix is here:
https://aomedia-review.googlesource.com/c/aom/+/193761

Added two test cases that trigger the failure.

Change-Id: I86490e3f96ffcea7fb6e972a7d2085675c65fcc4