rtc-svc: Fix to reset ref_idx for svc
Constrain the ref_idx reset only for single
spatial layers, since the frame_size written out
may be based on the refs, which can have a different
resolution. The previous fix of constraining to single
reference is not correct.

Another fix would be to set (after the reset)
cm->remapped_ref_idx[ref_frame - 1] = map_idx

For now taking this simpler fix of constraining
the reset to single spatial layers.

Bug: 369633254
Change-Id: I86490e3f96ffcea7fb6e972a7d2085675c65fcc4