Fix exit condition in rate correction update
The exit should only be done if the feature is enabled:
sf->overshoot_detection_cbr == FAST_DETECTION_MAXQ,
and on scene changes (rc->high_source_sad == 1).

The overshoot_detection feature is only enabled for
CBR, so for VBR this can cause the rate control not
to react to overshoot after scene change.

Change-Id: I585b4b3a2fa56ddf9e271b60be4b9abc3fdf842f