Introduce Adaptive CDEF
The current implementation works as an on/off switch. The encoder
only turns on CDEF when QP is >= 25 (qindex >= 100). This threshold
was chosen as CDEF benefits images in the low to medium quality
range, both by increasing SSIMU2 scores and subjective quality.

Future improvements to the mode may involve modulating CDEF strength
in a more granular way.

Bug: aomedia:375221136
Change-Id: I2b2279603e7f7ba8c4361728b9b4216263f948ca