rtc: Allow for color_sensitivity for speed 11
For speed 11 camera mode: when fixed partition
is being used, the color_sensitivity[] is unset (0)
and this leads to bad artifacts in color content.

Fix is to set a flag (x->fixed_partition_no_color_set) for
the superblock and use it the nonrd_pickmode to indicate
that block level color setting should be done.

This reduces artifacts, on rtc set:
bdrate gain ~2%, IC slowdown ~1.4%.

Change-Id: Id3663f8b5b82357d5fe8008a9f2430ba1a36c3b0