Introduce Variance Boost
Variance Boost is a variance AQ implementation that modulates
qindex depending on the ratio of low-variance to high-variance
8x8 subblocks within a 64x64 superblock, as well as the actual
variance of the subblocks themselves (the lower the variance,
the larger the boost).

This commit implements the qindex boosting and scaling formulas
that are optimized for still pictures.

Variance Boost is assigned `deltaq-mode=6`.

Bug: aomedia:375221136
Change-Id: I295b68eeb9f723addbef28054f4613bac3a24739