Introduce Variance Boost
Variance Boost is a variance AQ implementation that modulates
qindex depending on the ratio of low-variance to high-variance
8x8 subblocks within a 64x64 superblock, as well as the actual
variance of the subblocks themselves (the lower the variance,
the larger the boost).

This commit implements the qindex boosting and scaling formulas
that are optimized for still pictures.

Variance Boost is assigned `deltaq-mode=6`.

Approximate BD-Rate gains over deltaq-mode=2 sb-size=64 - cpu-used=6
(Daala's subset1):
- SSIMULACRA2 60: -4.7%
- SSIMULACRA2 70: -4.4%
- SSIMULACRA2 80: -2.7%
- SSIMULACRA2 90: -2.7%

Bug: aomedia:375221136
Change-Id: I295b68eeb9f723addbef28054f4613bac3a24739