encoder API: allow specifying layer-specific metadata
The AV1 specification Section 6.2.2 lists which OBUs are layer-specific
(must have the obu_extension_header when there are non-zero operating points)
and those which are not (must never have an obu_extension_header).
For OBU_METADATA, it depends on the type of metadata, as listed in the table
in Section 6.7.1. In particular, METADATA_TYPE_ITUT_T35 can be layer-specific
depending on the payload (so the user decides whether it's layer-specific).

This change adds a flag to allow setting a metadata as layer-specific.
This flag is added to aom_metadata_insert_flags.
Despite its name, this type seems to have been used more as an enum than
a bit field. With this change, the 4 lowest bits are used for the existing
metadata location values (key frame, inter frame, any frame), and the
AOM_MIF_LAYER_SPECIFIC uses the 5th bit.

One drawback is that when combining flags such as
(AOM_MIF_KEY_FRAME|AOM_MIF_LAYER_SPECIFIC) a cast to aom_metadata_insert_flags_t
is required since it's not defined as a uint32_t or similar.

In CWG-E050 there will also be new layer-specific metadata types introduced,
some of which are layer-specific, so the ability to write layer-specific
metadata in libaom will also be useful. In the case of CWG-E050, those
metadata types are always layer-specific so it could be hardcoded in the
library, but introducing the flag adds more flexibility.

Also fixes a bug where metadata was lost if the frame was resized.

Bug: 377851082

Change-Id: Ice49b9fa4ee3053d0eb93d6bec821dbac1bd153e