rtc-screen: Prune palette testing for speed 11
Allow for pruning the testing of the palette mode
to reduce spikes in encode_time, as observed in the
issue below. This is only done for speed 11 screen
content mode for now. The pruning targets content/blocks
where palette prediction is not expected to be useful,
so loss in quality is acceptable.

Also rename perc_flat_blocks_keyframe since that
variable is also computed for delta frames on
scene changes.

Bug: 380247338
Change-Id: I3a18d7c53e1c0209ed1fbce983c29775dcb14125