rtc-screen: Prune palette testing for speed 11
On scene changes for speed 11 screen mode:
prune the testing of the palette mode to reduce
spikes in encode_time, as observed in the issue below.

The pruning is done when the palette mode is not expected
to be so effective, so there is a reduction in the big
encode_time spikes, with minimal/acceptable quality loss.

Bug: 380247338
Change-Id: If32179ceb438922304315fc5260a11c58103660f