Variance Boost: 3-octile variance sample
Sample variance values at octiles 4, 5, and 6; then average them in
a 1:2:1 ratio to get a composite value which is then fed to the boost
formula.

This improves bit distribution within mixed-variance superblocks
(i.e. superblocks with both low-variance and high-variance subblocks).

This results in a -0.3% BD-Rate change in SSIMULACRA2 60-80 average,
-0.5% at 10 pct, and no change in SSIMULACRA2 90 (Daala's subset1).

Change-Id: Ibe4e41cd40eb4181172f1f53751a89d00eb80373